+++ lib/cucumber/filters/prepare_world.rb
@@ -24,9 +24,8 @@ module Cucumber
           end
           around_hooks = [init_scenario] + @original_test_case.around_hooks
 
-          default_hook = Cucumber::Hooks.before_hook(@original_test_case.source) do
-            #no op - legacy format adapter expects a before hooks
-          end
+          empty_hook = proc {} #no op - legacy format adapter expects a before hooks
+          default_hook = Cucumber::Hooks.before_hook(@original_test_case.source, Cucumber::Hooks.location(empty_hook), &empty_hook)
           steps = [default_hook] + @original_test_case.test_steps
 
           @original_test_case.with_around_hooks(around_hooks).with_steps(steps)