+++ lib/cucumber/runtime/after_hooks.rb
@@ -1,8 +1,9 @@
 module Cucumber
   class Runtime
     class AfterHooks
-      def initialize(action_blocks)
-        @action_blocks = action_blocks
+      def initialize(hooks, scenario)
+        @hooks = hooks
+        @scenario = scenario
       end
 
       def apply_to(test_case)
@@ -14,11 +15,11 @@ module Cucumber
       private
 
       def after_hooks(source)
-        @action_blocks.map do |action_block|
-          Hooks.after_hook(source, &action_block)
-        end
+        @hooks.map do |hook|
+          action = ->(result) { hook.invoke('After', @scenario.with_result(result)) }
+          Hooks.after_hook(source, Hooks.location(hook), &action)
+        end          
       end
     end
   end
 end
-