+++ lib/cucumber/runtime/before_hooks.rb
@@ -1,8 +1,11 @@
+require 'cucumber/hooks'
+
 module Cucumber
   class Runtime
     class BeforeHooks
-      def initialize(action_blocks)
-        @action_blocks = action_blocks
+      def initialize(hooks, scenario)
+        @hooks = hooks
+        @scenario = scenario
       end
 
       def apply_to(test_case)
@@ -14,8 +17,9 @@ module Cucumber
       private
 
       def before_hooks(source)
-        @action_blocks.map do |action_block|
-          Hooks.before_hook(source, &action_block)
+        @hooks.map do |hook|
+          action_block = ->(result) { hook.invoke('Before', @scenario.with_result(result)) }
+          Hooks.before_hook(source, Hooks.location(hook), &action_block)
         end
       end
     end